azure-diagnostic/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
azure-servicebus==7.12.0
azure-core==1.30.1
azure-monitor-opentelemetry==1.3.0